change the meaning of dry_run

+++ deployer/src/deployer/upload.py
@@ -119,9 +119,10 @@ class UploadFileTask(UploadTask):
     Class for file upload tasks.
     """
 
-    def __init__(self, file_path: Path, key: str):
+    def __init__(self, file_path: Path, key: str, dry_run=False):
         self.key = key
         self.file_path = file_path
+        self.dry_run = dry_run
 
     def __repr__(self):
         return f"UploadFileTask({self.file_path}, {self.key})"
@@ -201,16 +202,17 @@ class UploadFileTask(UploadTask):
         return f"max-age={cache_control_seconds}, public"
 
     def upload(self, bucket_manager):
-        bucket_manager.client.upload_file(
-            str(self.file_path),
-            bucket_manager.bucket_name,
-            self.key,
-            ExtraArgs={
-                "ACL": "public-read",
-                "ContentType": self.content_type,
-                "CacheControl": self.cache_control,
-            },
-        )
+        if not self.dry_run:
+            bucket_manager.client.upload_file(
+                str(self.file_path),
+                bucket_manager.bucket_name,
+                self.key,
+                ExtraArgs={
+                    "ACL": "public-read",
+                    "ContentType": self.content_type,
+                    "CacheControl": self.cache_control,
+                },
+            )
 
 
 class UploadRedirectTask(UploadTask):
@@ -220,9 +222,10 @@ class UploadRedirectTask(UploadTask):
 
     is_redirect = True
 
-    def __init__(self, redirect_from_key, redirect_to_key):
+    def __init__(self, redirect_from_key, redirect_to_key, dry_run=False):
         self.key = redirect_from_key
         self.to_key = redirect_to_key
+        self.dry_run = dry_run
 
     def __repr__(self):
         return f"UploadRedirectTask({self.key}, {self.to_key})"
@@ -235,14 +238,15 @@ class UploadRedirectTask(UploadTask):
         return f"max-age={HASHED_CACHE_CONTROL}, public"
 
     def upload(self, bucket_manager):
-        bucket_manager.client.put_object(
-            Body=b"",
-            Key=self.key,
-            ACL="public-read",
-            CacheControl=self.cache_control,
-            WebsiteRedirectLocation=self.to_key,
-            Bucket=bucket_manager.bucket_name,
-        )
+        if not self.dry_run:
+            bucket_manager.client.put_object(
+                Body=b"",
+                Key=self.key,
+                ACL="public-read",
+                CacheControl=self.cache_control,
+                WebsiteRedirectLocation=self.to_key,
+                Bucket=bucket_manager.bucket_name,
+            )
 
 
 class BucketManager:
@@ -300,7 +304,7 @@ class BucketManager:
                 break
         return result
 
-    def iter_file_tasks(self, build_directory, for_counting_only=False):
+    def iter_file_tasks(self, build_directory, for_counting_only=False, dry_run=False):
         # Prepare a computation of what the root /index.html file would be
         # called as a S3 key. Do this once so it becomes a quicker operation
         # later when we compare *each* generated key to see if it matches this.
@@ -329,9 +333,11 @@ class BucketManager:
             if for_counting_only:
                 yield 1
             else:
-                yield UploadFileTask(fp, key)
+                yield UploadFileTask(fp, key, dry_run=dry_run)
 
-    def iter_redirect_tasks(self, content_roots, for_counting_only=False):
+    def iter_redirect_tasks(
+        self, content_roots, for_counting_only=False, dry_run=False
+    ):
         # Walk the content roots and yield redirect upload tasks.
         for content_root in content_roots:
             # Look for "_redirects.txt" files up to two levels deep to
@@ -354,7 +360,8 @@ class BucketManager:
                             yield 1
                         else:
                             yield UploadRedirectTask(
-                                *self.get_redirect_keys(from_url, to_url)
+                                *self.get_redirect_keys(from_url, to_url),
+                                dry_run=dry_run,
                             )
 
     def count_file_tasks(self, build_directory):
@@ -369,18 +376,23 @@ class BucketManager:
         content_roots,
         existing_bucket_objects=None,
         on_task_complete=None,
+        skip_redirects=False,
+        dry_run=False,
     ):
         with concurrent.futures.ThreadPoolExecutor(
             max_workers=MAX_WORKERS_PARALLEL_UPLOADS
         ) as executor, StopWatch() as timer:
             # Upload the redirects first, then the built files. This
             # ensures that a built file overrides its stale redirect.
-            for task_iter in (
-                lambda: self.iter_redirect_tasks(content_roots),
-                lambda: self.iter_file_tasks(build_directory),
-            ):
+            task_iters = []
+            if not skip_redirects:
+                task_iters.append(
+                    self.iter_redirect_tasks(content_roots, dry_run=dry_run)
+                )
+            task_iters.append(self.iter_file_tasks(build_directory, dry_run=dry_run))
+            for task_iter in task_iters:
                 futures = {}
-                for task in task_iter():
+                for task in task_iter:
                     # Note: redirect upload tasks are never skipped.
                     if existing_bucket_objects and not task.is_redirect:
                         s3_obj = existing_bucket_objects.get(task.key)
@@ -451,23 +463,22 @@ def upload_content(build_directory, content_roots, config):
 
     totals = Totals()
 
-    if dry_run:
-        upload_timer = StopWatch()
-    else:
-        with DisplayProgress(
-            total_redirects + total_possible_files, show_progress_bar
-        ) as progress:
+    with DisplayProgress(
+        total_redirects + total_possible_files, show_progress_bar
+    ) as progress:
 
-            def on_task_complete(task):
-                progress.update(task)
-                totals.count(task)
+        def on_task_complete(task):
+            progress.update(task)
+            totals.count(task)
 
-            upload_timer = mgr.upload(
-                build_directory,
-                content_roots,
-                existing_bucket_objects,
-                on_task_complete=on_task_complete,
-            )
+        upload_timer = mgr.upload(
+            build_directory,
+            content_roots,
+            existing_bucket_objects,
+            on_task_complete=on_task_complete,
+            skip_redirects=not upload_redirects,
+            dry_run=dry_run,
+        )
 
     if dry_run:
         log.info("No uploads. Dry run!")
